Rewrite to repackage main jar artifact
Rewiting plugin to repackage main jar artifact as a plug. The behiviour
is similar to Spring Boot plugin, as source artifact can hold any number
of user customisations like custom files and manifests. Reusing it make
whole process dependant on artifact, which  is a minor downside, but
offers better coverage of all users needs.

To be able to easily test and also to enable to write other plugins in
future, like Gradle, I need to move code that is not directly linked to
maven to separate package: core.

 * Maven backed SLF4J Logger
 * Maven backed coordinates
 * Almost all of configuration, except Project class impl
 * Maven libraries scaffold
 * All Maven backend factories should be completed
 * Switched to VAVR on Core nad Maven tools
 * Cross platform fixes